+++ lib/src/tabs/keyed_tab_bar.dart
@@ -1,4 +1,3 @@
-import 'package:flutter/gestures.dart';
 import 'package:flutter/material.dart';
 
 import 'keyed_tab_controller.dart';
@@ -13,51 +12,34 @@ import 'keyed_tab_controller.dart';
 /// )
 /// ```
 class KeyedTabBar<K> extends TabBar {
+  ///
   KeyedTabBar({
-    Key? key,
+    super.key,
     required Map<K, Widget> tabs,
     required KeyedTabController<K> controller,
-    bool isScrollable = false,
-    EdgeInsetsGeometry? padding,
-    Color? indicatorColor,
-    bool automaticIndicatorColorAdjustment = true,
-    double indicatorWeight = 2.0,
-    EdgeInsetsGeometry indicatorPadding = EdgeInsets.zero,
-    Decoration? indicator,
-    TabBarIndicatorSize? indicatorSize,
-    Color? labelColor,
-    TextStyle? labelStyle,
-    EdgeInsetsGeometry? labelPadding,
-    Color? unselectedLabelColor,
-    TextStyle? unselectedLabelStyle,
-    DragStartBehavior dragStartBehavior = DragStartBehavior.start,
-    MaterialStateProperty<Color?>? overlayColor,
-    MouseCursor? mouseCursor,
-    bool? enableFeedback,
+    super.isScrollable,
+    super.padding,
+    super.indicatorColor,
+    super.automaticIndicatorColorAdjustment,
+    super.indicatorWeight,
+    super.indicatorPadding,
+    super.indicator,
+    super.indicatorSize,
+    super.labelColor,
+    super.labelStyle,
+    super.labelPadding,
+    super.unselectedLabelColor,
+    super.unselectedLabelStyle,
+    super.dragStartBehavior,
+    super.overlayColor,
+    super.mouseCursor,
+    super.enableFeedback,
     ValueChanged<K>? onTap,
-    ScrollPhysics? physics,
+    super.physics,
   }) : super(
-    key: key,
-    tabs: controller.mapToList(tabs),
-    controller: controller,
-    isScrollable: isScrollable,
-    padding: padding,
-    indicatorColor: indicatorColor,
-    automaticIndicatorColorAdjustment: automaticIndicatorColorAdjustment,
-    indicatorWeight: indicatorWeight,
-    indicatorPadding: indicatorPadding,
-    indicator: indicator,
-    indicatorSize: indicatorSize,
-    labelColor: labelColor,
-    labelStyle: labelStyle,
-    labelPadding: labelPadding,
-    unselectedLabelColor: unselectedLabelColor,
-    unselectedLabelStyle: unselectedLabelStyle,
-    dragStartBehavior: dragStartBehavior,
-    overlayColor: overlayColor,
-    mouseCursor: mouseCursor,
-    enableFeedback: enableFeedback,
-    onTap: onTap == null ? null : (index) => onTap(controller.keys[index]),
-    physics: physics,
-  );
+          tabs: controller.mapToList(tabs),
+          controller: controller,
+          onTap:
+              onTap == null ? null : (index) => onTap(controller.keys[index]),
+        );
 }